Update endpoints, move test files
