+++ src/tests/actions/postLogInInfoAction.test.js
@@ -1,9 +1,9 @@
 import moxios from 'moxios'
 import thunk from 'redux-thunk'
 import configureMockStore from 'redux-mock-store'
-import { postLogInInfo } from '../actions/postLogInInfoAction'
-import { POST_LOGIN_INFO } from '../types'
-import logInResponse from '../fixtures/logIn'
+import { postLogInInfo } from '../../actions/postLogInInfoAction'
+import { POST_LOGIN_INFO } from '../../types'
+import logInResponse from '../../fixtures/logIn'
 
 const middlewares = [thunk]
 const mockStore = configureMockStore(middlewares)
@@ -23,7 +23,7 @@ describe('postLogInInfo action', () => {
     const expectedActions = [
       { type: POST_LOGIN_INFO, payload: logInResponse }
     ]
-    moxios.stubRequest('http://localhost:3000/users/sign_in', {
+    moxios.stubRequest('/users/sign_in', {
       status: 200,
       response: logInResponse
     })
