--- src/tests/store.test.js
@@ -1,25 +1,0 @@
-import store from '../store'
-
-describe('Store', () => {
-  it('should update', async () => {
-    await store.dispatch({
-      type: 'GET_USERS',
-      payload: ['Run the tests']
-    })
-
-    expect(store.getState()).toEqual({
-      users: ['Run the tests']
-    })
-  })
-
-  it('should update after login', async () => {
-    await store.dispatch({
-      type: 'POST_LOGIN_INFO',
-      payload: ['Rerun them']
-    })
-
-    expect(store.getState()).toEqual({
-      users: ['Rerun them']
-    })
-  })
-})
